+++ src/app/page.tsx
@@ -5,14 +5,17 @@ import { redirect } from 'next/navigation';
 export default async function Page({
   searchParams,
 }: {
-  searchParams?: { callbackUrl?: string };
+  searchParams?: Promise<Record<string, string | string[] | undefined>>;
 }) {
   const cookieStore = await cookies();
   const hasToken = !!cookieStore.get(
     process.env.JWT_COOKIE_NAME || 'devseed_token',
   )?.value;
+  const sp = (await searchParams) ?? {};
+  const cb = typeof sp.callbackUrl === 'string' ? sp.callbackUrl : '/home';
+
   if (hasToken) {
-    redirect(searchParams?.callbackUrl || '/home');
+    redirect(cb);
   }
 
   return <OnboardingView />;
